Rename collections package to structures
