+++ src/main/java/views/FacultyView.java
@@ -1,6 +1,6 @@
 package main.java.views;
 
-import main.java.collections.List;
+import main.java.structures.List;
 import main.java.services.FacultyService;
 import main.java.models.Course;
